Refactor the call method to core
